Add states and transitions between town and arena
