+++ src/maps/arena.js
@@ -27,11 +27,15 @@ export default {
 
     // Collide with player
     map.setCollisionBetween(1, 10000, true, layer2);
+
+    // Enter town when player collides with exit layer
+    map.setTileLocationCallback(1, 14, 1, 2, () => (game.state.start('town')), this, exitLayer);
   },
   update: (game, collidables) => {
     collidables.forEach(item => {
       // Collide with item
       game.physics.arcade.collide(item, layer2);
+      game.physics.arcade.collide(item, exitLayer);
     });
   }
 };
